Modify ingress (#25)
Add support for modify ingress. Actually implemented by creating a new one, and then deleting the old one.

Prior naming of ingress caused conflicts during creation (since an ingress with that name already exist - because we haven't deleted the old one yet).
Initially we attempted to include a random part in the ingress name to remedy this. However, this was problematic in ServiceWatcher, where we need to delete an ingress based on a TenantConfig. Therefore, I set the ingress name so that it includes a SHA1 hash of all relevant values instead.

Co-authored-by: Rickard Andersson <rickard.andersson@devies.se>
Co-authored-by: Fredrik Mile <fredrik.mile@gmail.com>