Modify ingresses on service port change (#28)
When the port of a service is changed, a modify event will be triggered. Delete old ingress and create a new one with the new port.

In our e2e test we now use the -jsonpath argument to kubectl wait which requires a later a version of kubectl than what is currently included in the ubuntu-latest image. Hence, we now set both kind version and kubectl_version with the setup-kind action. Note that the kubectl_version argument has not been released yet, therefore we use the current latest commit to main branch as version.